fix 'list' verbose for controller namespaces
Checking that namespace against the subsystem's namespace list is not
always correct because the namespace may not be multipath capable. For
those namespaces, the kernel doesn't link the namespace to the
subsystem, so it ends up not getting displayed. Print all the
controller's namespaces, as the topology should have filtered ones out
that the controller is not a path to.

Signed-off-by: Keith Busch <kbusch@kernel.org>