fabrics: correctly handle ctrl_loss_tmo settings for loop
The previous fix had an issue with referrals, as it would take
the default values and apply them to all referral entries.
And if the default entries were for 'loop', the default ctrl_loss_tmo
setting would not be used, but rather '-1'.
So this patch reverts the previous patch and correctly blanks
out the ctrl_loss_tmo setting when constructing the connect string.

Fixes: bdf4f3b ("fabrics: ctrl_loss_tmo setting is invalid for 'loop'")
Signed-off-by: Hannes Reinecke <hare@suse.de>