tree: error checking in nvme_init_ctrl()
nvme_init_ctrl() is called to initialize a controller after a
'connect' call. As this might fail for a variety of reasons we
need to ensure to change the controller state if we need to
return an error; this ensures that we can repeat a call to
nvme_init_ctrl() after the error causing the failure has been
resolved.

Signed-off-by: Hannes Reinecke <hare@suse.de>