nvme: open namespace device in exclusive mode
To prevent from accidently format a namespace which is still in use
the namespace device gets opend exclusivly. If the device is in use~$
this will fail with the error EBUSY. When nvme is called with the force$
commandline option it will format the namespace even when it's in use.

Signed-off-by: Olaf Schmerse <olaf.schmerse@ionos.com>