Unify configuration includes, use meson-generated configuration data
Currently, we have two separate configuration includes: ccan/config.h
and config-host.h. The first is generated by the ccan configurator, the
second by the configuration step (either ./configure, or meson).

This change unifies these into a top-level config.h include.

For the meson case, we can use the meson configuration file to provide
all data. For the configure case, we still use the ccan configurator,
but include the resulting file from the top-level config.h

This allows us to cross-compile with meson.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>