nvme: fail on busy namespace
To avoid formarting a busy namespace accidently a new function called
open_exclusive() is implemented. Instead of opening it always it opens
the device file in exclusive mode (O_EXCL). If this fails with the error EBUSY an error
messages hints to the --force command line option. By calling nvme format with the --force option it
will continue formating the namespace.

This patch fixes issue: #1095
Signed-off-by: Olaf Schmerse <olaf.schmerse@ionos.com>