meson: use library() and rely on user settings
There is no need to have both _static and _shared targets. Just use
library(). Project using libnvme as a subproject can now just set

  default_options: ['default_library=static']

in the subproject statement to choose the library version.

For pkgconfig, simplify and let the module figure out Libs and Requires.
This has the effect of removing Requires and making it a
Requires.private so we don't require libnvme consumers to link with
libuuid and json-c (unless they specifically needs to use it like
test/test.c).

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>