Make Python Bindings optional
Added a new configuration parameter "python". By default it is set to 'auto',
which means that the Python bindings will be built if all dependencies can be
found on the host. If not, Python bindings won;t be built. The option can
also be set to 'true' or 'false' to force Python bindings to be built or not.

Signed-off-by: Martin Belanger <martin.belanger@dell.com>