build: Install additional files incl documentation
Install additional files such as dracut, systemd and udev scipts.
Also add support for building and installing documentation.

Introduce a build option '-Ddocs=[all,man,html]' which builds (if
asciidoctor is availalbe) and installs the docs.

meson doesn't support pre compiled sources updated in the source
tree. This is a design decision. That means the man pages
(Documentation/*.1) and the html pages (Documentation/*.html) can't be
updated so easily as with Makefile.

The workaround provide a shell script which updates the in tree
pre compiled documentation

Signed-off-by: Daniel Wagner <dwagner@suse.de>