tree: scan controllers before scanning subsystems in nvme_scan_topology()
The linux sysfs hierarchy only knows about 'controllers' and 'subsystems',
so when scanning the topology we should follow this structure and scan
first for controllers and then for subsystems.
Just scanning for subsystems and all controllers attached to them has
the problem that the hostnqn can only be derived from the controller,
but by the time the controller is scanned the subsystem has already
been created and attached to the default host.
Scanning for controllers first will setup the correct hosts
(and subsystems), so the subsequent scan for subsystems is primarily
there to find all namespaces attached to the subsystems.

Signed-off-by: Hannes Reinecke <hare@suse.de>