build: Rename config.h to libnvme-config.h
The meson documentation recommends not to put a config.h file in the
external search path. This will clash with any project consuming this
library. There are two ways to workaround, either renaming the config
header file or put it under a different path which isn't added into
the default include path.

Renaming is way simpler, let's do this.

Signed-off-by: Daniel Wagner <dwagner@suse.de>