build: Do not use the configurator for meson
meson does the feature detection itself and we don't have to rely on
the configurator. So don't use it.

To avoid touching the ccan files, we have to rethink our config.h to
libnvme-config.h rename change. Since the ccan files include
'config.h' we have to use the second option to avoid including it into
external include paths. This is moving the config.h file into a
private folder and drop the include path '.'.

Signed-off-by: Daniel Wagner <dwagner@suse.de>