build: Set maximum warning level
Without defining the default warning level the build system
will use the default level, which is implementation depended.
For example muon sets it to 3 which includes '-Wpendantic'.
This results in a lot of ISO-C non compliant warnings.

Let's define the warning level so that all build systems
are using the same values.

Signed-off-by: Daniel Wagner <dwagner@suse.de>