nvme: Return 0 on success and 1 on error
main() should return 0 on success and 1 on error. Generally there are
not fixed rules. Though SUSv3 specifies EXIT_SUCCESS(0) and
EXIT_FAILURE(1).

The main reason why we should avoid anything but 0/1 is the shell
interpretes those return codes. Values above 128 can cause confusion
in shell scripts. When a command is terminated by a signal, the
shell indicates this fact by setting the value of the variable $?
to 128 plus the signal number.

See also
  https://stackoverflow.com/questions/1101957/are-there-any-standard-exit-status-codes-in-linux
  https://tldp.org/LDP/abs/html/exitcodes.html

Signed-off-by: Daniel Wagner <dwagner@suse.de>