nvme: Adapt to logging changes in libnvme
libnvme allows to define which file descriptor should be used for
logging. This is only for the fabric related interface relevant hence
this is attached to the nvme_root_r object.

Use common code for this into a helper which maps the verbose command
line levels to the correct log levels.

Signed-off-by: Daniel Wagner <dwagner@suse.de>