nvme: Only open device once
open_exclusive() takes care with opening the device. Do not open a
second time the device.

open_exclusive() has an overwrite flag. Document it a bit better by
renaming it what it actually means. The idea here is that the device
is openened with O_EXCL and when forced is used, open without this
flag.

Signed-off-by: Daniel Wagner <dwagner@suse.de>