netapp-nvme: fix nvme ns desc uuid handling for ontapdevices
The nvme ns desc uuid handling in the NetApp plugin for
ontapdevices is broken after the recent upgrade to libnvme. The
uuid hanlding logic here was flawed to begin with, so rework this
code to ensure the respective uuid descriptor itself is stored in
the ontapdevice_info structure and processed appropriately.

Signed-off-by: Martin George <marting@netapp.com>