nvme: Print full device path
Restore nvme-cli 1.x behavior to print the full path to the device
file. We assume that the devtmpfs is mounted under /dev which is the
case for any systemd based distro and a de-facto standard.

But SPDK users see their device populated under /dev/spdk have their
device file appear. So instead blindly prefix /dev do a stat to check
if the file exist.

If our educated guesses fail, don't add the full path. This should be
really a corner case and as long no one complains don't overengineer.

Signed-off-by: Daniel Wagner <dwagner@suse.de>