fabrics: Avoid out of bounds string chomping
Update all callsides of nvme_chomp to use the correct max lenght of
the string to avoid out of bounds access.

While at it, nvme_chomp() is used by fabrics.c to sanitize the traddr
and trsvcid. There is no user outside of fabrics and clearly this is a
small helper function. Thus move it to fabrics and rename it to a more
comon name strchomp.

Signed-off-by: Daniel Wagner <dwagner@suse.de>