fabrics: Introduce force flag to overwrite persistence logic
The persistence flag will be overwritten by the built in logic. Allow
the user to overwrite this by providing a force command line option.

While at it make the 'discover_from_config_file' path behave as the
'nvmf_discover' path. The idea here is that whenever a matching
persistent discovery controller is present just use this one. Only
create a new discovery controller when none matching found or
we are asked to do so via --force.

Signed-off-by: Daniel Wagner <dwagner@suse.de>