fabrics: Invoke nvmf_dim() with provided tas argument
Looks like a copy-paste issue. One of the arguments to
nvmf_register_ctrl() is "tas" (i.e. the DIM Task). This argument
should be passed to the nvmf_dim() API, but instead a fixed value
of NVMF_DIM_TAS_REGISTER was passed.

The "tas" field specifies whether to perform a "registration",
"deregistration", or a "registration update".

Signed-off-by: Martin Belanger <martin.belanger@dell.com>