tree: Remove default port setting for TCP and RDMA ports
When the controller is created, the discovery_ctrl attribute is not
set yet. The caller is supposed to set this after the creation of the
new discovery controller. Thus we can't really set the default port at
this level. This is a bit of a chicken-egg situation.

Move this work up to the caller to provide the correct defaults.

Signed-off-by: Daniel Wagner <dwagner@suse.de>