mi: Add basic tests
This change adds a couple of basic tests for the mi infrastrcucture.

To do this, we define a test "transport", which just allows inserting a
custom function as the transport's ->submit callback. Each test is then
a pair of: the test itself, and a callback to implement a specific
response from the test "device".

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>