ci: publish only sdist on PyPi
Pre building the binaries with wheels is difficult, as we would also
need to ship the libnvme.so with the Python binding.

This is a well known limitation of this kind of setup (binding to a
shared library). Most project fallback to just ship the sdist.

Though there is a drawback as the C library is missing and the user
has to provide the library himself, with all problems which come along
with setup. But it seems common practice with other Python bindings,
so we don't want to be hostile to the ones which know what they are
doing.

Normal users should just use the distribution packages anyway.

Signed-off-by: Daniel Wagner <dwagner@suse.de>